requirements.txt
Django>=4,<5
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
dj-database-url
django-cors-headers
